docker: Force Unix line endings on docker/ files
When using git on Windows, git may try to "fix" line endings to CRLF.
Then, when building the Docker image, the files copied can end up having
the wrong line ending, which causes scripts to fail to run.

This patch fixes the problem by using .gitattributes to indicate to git
which line ending to use for the files in the docker/ directory.

https://github.com/albertito/chasquid/pull/66

Amended-by: Alberto Bertogli <albertito@blitiri.com.ar>
  Adjusted commit message, extended comment on .gitattributes.